feat: skip authorization for followup commands
Followup commands are written by Zeebe. These do not have any authorizations on them, so we cannot check them. We also
do not need to check them, Zeebe is authorized to do anything.

We can identify these commands by looking at the request metadata. The request metadata indicates if there's a
requestId, or a requestStreamId. If one of these is available we know that this request is coming from a client.